+++ engine/src/graphics/vertex.cpp
@@ -4,7 +4,7 @@
 
 void liminal::vertex::add_bone_data(unsigned int id, float weight)
 {
-    for (unsigned int i = 0; i < NUM_BONES_PER_VERTEX; i++)
+    for (size_t i = 0; i < NUM_BONES_PER_VERTEX; i++)
     {
         if (bone_ids[i] == NUM_BONES_PER_VERTEX)
         {
